Extract commands from being inlined
As there is quite a lot of code in the extension.ts file and it would be
nice to break it out to separate chunks if possible.
