+++ src/names.ts
@@ -1,0 +1,3 @@
+export function toCIdentifier(name: string) {
+    return name.toLowerCase().replace(/[@,-]/g, '_').replace(/[#&]/g, '');
+}
